@@ -1,29 +1,46 @@
-import React, { useState } from 'react';
-import EmployeeList from './EmployeeList';
-import EmployeeForm from './EmployeeForm';
+import React, { useContext } from "react";
+import {
+  BrowserRouter,
+  Routes,
+  Route,
+  Navigate,
+} from "react-router-dom";
 
-function App() {
-    const [selectedEmployee, setSelectedEmployee] = useState(null);
-    const [addingNew, setAddingNew] = useState(false);
+import { AuthContext } from "./AuthContext";
+import Login from "./Login";
+import Register from "./Register";
+import EmployeeList from "./EmployeeList";
+import EmployeeForm from "./EmployeeForm";
 
-    const handleFormClose = () => {
-        setSelectedEmployee(null);
-        setAddingNew(false);
-    };
+/* ----- Guards ---------------------------------------------------- */
+const PrivateRoute = ({ children }) => {
+  const { token } = useContext(AuthContext);
+  return token ? children : <Navigate to="/" replace />;
+};
+const AuthWrapper = ({ children }) => {
+  const { token } = useContext(AuthContext);
+  return token ? <Navigate to="/employees" replace /> : children;
+};
 
-    return (
-        <div className="App">
-            <h1>Employee Directory</h1>
-            <button onClick={() => setAddingNew(true)}>Add Employee</button>
-            <EmployeeList onSelectEmployee={setSelectedEmployee} />
-            {(selectedEmployee || addingNew) && (
-                <EmployeeForm
-                    employee={selectedEmployee}
-                    onUpdate={handleFormClose}
-                />
-            )}
-        </div>
-    );
+/* ----- Router ---------------------------------------------------- */
+export default function App() {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<AuthWrapper><Login /></AuthWrapper>} />
+        <Route path="/register" element={<AuthWrapper><Register /></AuthWrapper>} />
+
+        <Route
+          path="/employees"
+          element={<PrivateRoute><EmployeeList /></PrivateRoute>}
+        />
+        <Route
+          path="/employees/add"
+          element={<PrivateRoute><EmployeeForm /></PrivateRoute>}
+        />
+
+        <Route path="*" element={<Navigate to="/" replace />} />
+      </Routes>
+    </BrowserRouter>
+  );
 }
-
-export default App;
